Module 18 - Animations
This module was about animations, the class videos were for beginners and didnt involve C++, so I tried to add things that were related to animations and C++, like the anim notifies and a combo system, that can be added just by adding more animations into the anim montages array of the ability. Also made myself the animations using the ControlRig which I didn't have experience using yet.